A few minor improvements (#59)
* Make the docstrings more consistent

* Remove unused function argument default value

Remove the default value for the "variables" argument of the DeltaPy
constructor. The DeltaPy constructor is always called with a
"variables" argument so there is no need for a default value.

* Use {} rather than dict() to create an empty dict

It is recommended to use {}, [] and () to create an empty dict, list and
tuple, respectively. In particular, {}, [] and () are known to be faster.